refactor: update ui text
