+++ app/src/main/java/com/example/eventapp/ui/profiles/AdminProfilesFragment.java
@@ -1,4 +1,4 @@
-package com.example.eventapp.ui.dashboard;
+package com.example.eventapp.ui.profiles;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,22 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.eventapp.databinding.FragmentDashboardBinding;
+import com.example.eventapp.databinding.FragmentAdminProfilesBinding;
 
-public class DashboardFragment extends Fragment {
+public class AdminProfilesFragment extends Fragment {
 
-    private FragmentDashboardBinding binding;
+    private FragmentAdminProfilesBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        DashboardViewModel dashboardViewModel =
-                new ViewModelProvider(this).get(DashboardViewModel.class);
+        AdminProfilesViewModel adminProfilesViewModel =
+                new ViewModelProvider(this).get(AdminProfilesViewModel.class);
 
-        binding = FragmentDashboardBinding.inflate(inflater, container, false);
+        binding = FragmentAdminProfilesBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textDashboard;
-        dashboardViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textProfiles;
+        adminProfilesViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
