+++ app/src/main/java/com/example/eventapp/ui/scanqr/ScanQrFragment.java
@@ -1,4 +1,4 @@
-package com.example.eventapp.ui.notifications;
+package com.example.eventapp.ui.scanqr;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,22 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.eventapp.databinding.FragmentNotificationsBinding;
+import com.example.eventapp.databinding.FragmentScanQrBinding;
 
-public class NotificationsFragment extends Fragment {
+public class ScanQrFragment extends Fragment {
 
-    private FragmentNotificationsBinding binding;
+    private FragmentScanQrBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        NotificationsViewModel notificationsViewModel =
-                new ViewModelProvider(this).get(NotificationsViewModel.class);
+        ScanQrViewModel scanQrViewModel =
+                new ViewModelProvider(this).get(ScanQrViewModel.class);
 
-        binding = FragmentNotificationsBinding.inflate(inflater, container, false);
+        binding = FragmentScanQrBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textNotifications;
-        notificationsViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textScanQr;
+        scanQrViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
