+++ app/src/main/res/layout/fragment_scan_qr.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.dashboard.DashboardFragment">
+    tools:context=".ui.scanqr.ScanQrFragment">
 
     <TextView
-        android:id="@+id/text_dashboard"
+        android:id="@+id/text_scan_qr"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
