+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -13,13 +13,13 @@
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.example.eventapp.ui.dashboard.DashboardFragment"
-        android:label="@string/title_dashboard"
-        tools:layout="@layout/fragment_dashboard" />
+        android:name="com.example.eventapp.ui.scanqr.ScanQrFragment"
+        android:label="@string/title_scanqr"
+        tools:layout="@layout/fragment_scan_qr" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.example.eventapp.ui.notifications.NotificationsFragment"
-        android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        android:name="com.example.eventapp.ui.profiles.AdminProfilesFragment"
+        android:label="@string/title_profiles"
+        tools:layout="@layout/fragment_admin_profiles" />
 </navigation>
